Renaming emotion view classes
