+++ iOS/UI/Swipeable Input/EmotionView.swift
@@ -1,112 +1,61 @@
 //
-//  EmotionView.swift
+//  SwiftUIView.swift
 //  Jibber
 //
-//  Created by Benji Dodgson on 12/13/21.
+//  Created by Benji Dodgson on 12/16/21.
 //  Copyright © 2021 Benjamin Dodgson. All rights reserved.
 //
 
-import Foundation
-import UIKit
+import SwiftUI
 import StreamChat
 
-class EmotionView: BaseView {
+struct EmojiContainer: View {
+    @State var emoji: String = ""
     
-    let emojiLabel = ThemeLabel(font: .reactionEmoji)
-    
-    let label = ThemeLabel(font: .small)
-    let button = ThemeButton()
-    
-    var didSelectEmotion: ((Emotion) -> Void)?
-    
-    override func initializeSubviews() {
-        super.initializeSubviews()
-        
-        self.addSubview(self.emojiLabel)
-        
-        self.addSubview(self.label)
-        self.addSubview(self.button)
-        
-        self.clipsToBounds = false
-        
-        self.button.showsMenuAsPrimaryAction = true
+    var body: some View {
+        RoundedRectangle(cornerRadius: Theme.innerCornerRadius)
+            .overlay (
+                ZStack {
+                    RoundedRectangle(cornerRadius: Theme.innerCornerRadius)
+                        .stroke(.D6withAlpha, alpha: 0.3, width: 0.25)
+                    Text(self.emoji).fontType(.small)
+                }
+            ).frame(width: 20, height: 20, alignment: .center)
+            .color(.T1, alpha: 0.1)
     }
+}
+
+extension Binding {
+    func onChange(_ handler: @escaping (Value) -> Void) -> Binding<Value> {
+        return Binding(
+            get: { self.wrappedValue },
+            set: { selection in
+                self.wrappedValue = selection
+                handler(selection)
+        })
+    }
+}
+
+struct EmotionView: View {
     
-    func configure(for message: Messageable) {
-        let controller = ChatClient.shared.messageController(for: message)
+    let emotion: Emotion
         
-        guard let data = controller?.message?.extraData["emotions"] else {
-            self.isVisible = false
-            return
+    var body: some View {
+        HStack {
+            Spacer.length(.short)
+            EmojiContainer(emoji: self.emotion.emoji)
+            Spacer.length(.short)
+            Text(self.emotion.rawValue.firstCapitalized)
+                .fontType(.small)
+            Spacer.length(.short)
         }
+    }
+}
 
-        guard case .array(let JSONObjects) = data, let emotionJSON = JSONObjects.first else {
-            self.isVisible = false
-            return
+struct EmotionView_Previews: PreviewProvider {
+    static var previews: some View {
+        Group {
+            EmotionView(emotion: .awkward)
         }
-
-        guard case .string(let emotionString) = emotionJSON,
-              let emotion = Emotion(rawValue: emotionString) else {
-
-                  self.isVisible = false
-                  return
-              }
-
-        self.isVisible = true
-        self.configure(for: emotion)
-    }
-    
-    func configure(for emotion: Emotion) {
-        self.emojiLabel.setText(emotion.emoji)
-        self.label.setText(emotion.rawValue.firstCapitalized)
-        self.button.menu = self.createMenu(for: emotion)
-        self.setNeedsLayout()
-    }
-    
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        
-        self.emojiLabel.setSize(withWidth: 200)
-        self.label.setSize(withWidth: 200)
-        
-        self.emojiLabel.pin(.left)
-        
-        self.height = old_MessageDetailView.height
-        self.width = self.emojiLabel.width + Theme.ContentOffset.short.value + self.label.width + Theme.ContentOffset.short.value.doubled
-        
-        self.pin(.left)
-        
-        self.label.centerOnY()
-        self.label.match(.left, to: .right, of: self.emojiLabel, offset: .standard)
-        self.emojiLabel.center.y = self.label.center.y
-
-        self.button.expandToSuperviewWidth()
-        self.button.height = 36
-        self.button.centerOnXAndY()
-    }
-    
-    private func createMenu(for emotion: Emotion) -> UIMenu {
-        
-        var children: [UIMenuElement] = []
-        Emotion.allCases.forEach { e in
-            let state: UIMenuElement.State = e == emotion ? .on : .off
-            let title = "\(e.emoji) \(e.rawValue.capitalized)"
-            let action = UIAction(title: title,
-                                  image: nil,
-                                  identifier: nil,
-                                  discoverabilityTitle: nil,
-                                  attributes: [],
-                                  state: state) { [unowned self] _ in
-                self.didSelectEmotion?(e)
-                self.configure(for: e)
-            }
-            children.append(action)
-        }
-        
-        return UIMenu(title: "I'm feeling...",
-                      image: nil,
-                      identifier: nil,
-                      options: [.singleSelection],
-                      children: children)
     }
 }
